Refactor: Move get-system-instructions to its own module
The 'get-system-instructions' command was previously defined in 'src/mcp/utils.ts', which is not consistent with how other tools are defined.

This change moves the command to its own file, 'src/mcp/tools/system-tools.ts', and registers it in 'src/mcp/index.ts'. This improves the structure and maintainability of the codebase.
