+++ src/mcp/utils.ts
@@ -21,23 +21,6 @@ export let mcpServer = new McpServer({
     instructions: "The Phaser Editor MCP server exposes a set of tools for automating and extending Phaser Editor v5 through the Model Context Protocol (MCP). To use these tools, ensure Phaser Editor is running and accessible. This server enables integration with LLMs and external clients, allowing you to query, modify, and automate game scenes, assets, tilemaps, and more directly from your MCP-compatible environment.",
 });
 
-mcpServer.tool("get-system-instructions", "Get the system instructions. This is fully required for the LLM to know how to build the arguments of this MCP server tools. Tools like `scene-add-game-objects`, `scene-update-game-objects`, `scene-add-game-object-filters`, `scene-update-game-object-filters`, `scene-add-plain-objects`, and `scene-update-plain-objects` requires that the LLM get the system instructions first to know the tool arguments structure.", () => {
-
-    const __filename = fileURLToPath(import.meta.url);
-
-
-    const text1 = readFileSync(
-        join(dirname(__filename), "tools/prompts/system.md"), "utf-8");
-
-    const text2 = readFileSync(
-        join(dirname(__filename), "tools/prompts/tools.md"), "utf-8");
-
-    return {
-        content: [
-            { type: "text", text: `${text1}\n${text2}` }
-        ],
-    }
-});
 
 export async function startServer() {
